clipboard-apis/clipboard-file-manual.html

<!doctype html>
<meta charset="utf-8">
<title>Clipboard: DataTransfer File manual test</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#to-fire-a-clipboard-event">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/copy-data-transfer.js"></script>
<style>
#pastewrapper {
  display: block;
  width: 400px;
  height: 200px;
  position: relative;
  padding: 50px 0 0 100px;
}
#pastezone {
  display: block;
  border: 1px solid black;
  width: 200px;
  height: 100px;
}
</style>
<p>
  Please download <a download href="resources/copied-file.txt">this file</a>,
  and copy and paste it into the box below.
</p>

<div id="pastewrapper">
  <div id="pastezone">
    Paste Here
  </div>
</div>

<script>
'use strict';

const pasteWrapper = document.querySelector('#pastewrapper');
const pasteZone = document.querySelector('#pastezone');

const pastePromise = new Promise((resolve, reject) => {
  pasteZone.onpaste = async event => {
    event.preventDefault();
    resolve(copyDataTransfer(event.clipboardData));
  };
});

promise_test(() => {
  return pastePromise.then((dataTransfer) => {
    assert_true(dataTransfer.types.includes('Files'));
  });
}, 'DataTransfer.types in paste file');

promise_test(() => {
  return pastePromise.then((dataTransfer) => {
    assert_equals(
        dataTransfer.files.length, 1,
        'DataTransfer.files should have one element');
    const file = dataTransfer.files[0];
    assert_equals(
        file.error, undefined,
        'Reading the file in DataTransfer.files[0] should not throw');
    assert_true(
        file.file instanceof File,
        'DataTransfer.files[0] should be a File instance');
    assert_equals(file.file.name, 'copied-file.txt');
    assert_equals(file.file.type, 'text/plain');
    assert_equals(file.file.size, 21);
    assert_equals(file.data, 'The test has FAILED.\n');
  });
}, 'DataTransfer.files in paste');

promise_test(() => {
  return pastePromise.then((dataTransfer) => {
    const items = dataTransfer.items.filter(i => i.kind === 'file');
    assert_equals(items.length, 1,
        'DataTransfer.items[kind="file"] should have 1 element');
    const item = items[0];
    assert_equals(
        item.error, undefined,
        'Reading the file in DataTransfer.items[0] should not throw');
    assert_true(
        item.file instanceof File,
        'DataTransfer.items[0] should be a File instance');
    assert_equals(item.file.name, 'copied-file.txt');
    assert_equals(item.file.type, 'text/plain');
    assert_equals(item.file.size, 21);
    assert_equals(item.data, 'The test has FAILED.\n');
  });
}, 'DataTransfer.items in paste');
</script>
